Continue refactor changes
