+++ scala/opentransit/src/main/scala/com/azavea/opentransit/indicators/calculators/InterstopDistance.scala
@@ -1,7 +1,8 @@
-package com.azavea.opentransit.indicators
+package com.azavea.opentransit.indicators.calculators
 
 import com.azavea.gtfs._
 import com.azavea.opentransit._
+import com.azavea.opentransit.indicators._
 
 // This indicator calculates the average length of transit
 // line between two stops
