+++ scala/opentransit/src/main/scala/com/azavea/opentransit/indicators/calculators/TransitDensity.scala
@@ -1,4 +1,4 @@
-package com.azavea.opentransit.indicators
+package com.azavea.opentransit.indicators.calculators
 
 import com.azavea.gtfs._
 import com.azavea.opentransit._
