+++ scala/opentransit/src/main/scala/com/azavea/opentransit/indicators/calculators/TravelTimePerformance.scala
@@ -1,9 +1,10 @@
-package com.azavea.opentransit.indicators
+package com.azavea.opentransit.indicators.calculators
 
 import org.joda.time.Seconds
 
 import com.azavea.gtfs._
 import com.azavea.opentransit._
+import com.azavea.opentransit.indicators._
 import com.azavea.opentransit.indicators.parameters._
 
 /**
